# storage/api-docs/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice Handler API",
        "description": "Invoice Swagger API",
        "version": "0.1"
    },
    "paths": {
        "/api/invoices/store": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Create a new invoice",
                "operationId": "cf6c53e8552e92e08b7cde570b592070",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "description": "Status of the Invoice",
                                        "type": "string",
                                        "default": "draft",
                                        "enum": [
                                            "draft",
                                            "sending",
                                            "sent-to-client"
                                        ],
                                        "example": "draft"
                                    },
                                    "customer_name": {
                                        "description": "Customer Name",
                                        "type": "string",
                                        "example": "Customer Name"
                                    },
                                    "customer_email": {
                                        "description": "Customer Email",
                                        "type": "email",
                                        "example": "Customer Email"
                                    },
                                    "product_lines": {
                                        "description": "JSON data of product lines",
                                        "type": "string",
                                        "example": "{}"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice created successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/invoices/{id}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "View Invoice Information",
                "operationId": "53e4311d81b9d387cef639a96cff3bf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invoice Information"
                    },
                    "400": {
                        "description": "Invoice was not found"
                    }
                }
            }
        },
        "/api/invoices/send": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send Invoice",
                "operationId": "fe05b0ca1dcaa736c4a7925ef8c873f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Invoice has been sent successfully"
                    },
                    "400": {
                        "description": "Sending error"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Invoices",
            "description": "Invoice Management Endpoints"
        },
        {
            "name": "Notifications",
            "description": "Notification Management Endpoints"
        }
    ]
}